11/26 학습 일지
발사체 발사, 폭발, 탄종 선택, 발사 이동, 발사체 소멸(탄착과 수명 소멸 중 수명만 구현).

옵션 창을 게임패널에서 실행 가능하도록 구현함.(단 옵션창 선택시 게임루프 정지 로직 구현 안함. 수정 필요) 이후 이미지
버튼으로 기존 게임 창의 버튼 UI 변경할 예정.

배율 변경: 기존 이미지 1/2 배율에서, 1/3로 변경.
---------------------------------------
이후 추가할 내용
1. 충돌 로직 완성하기
2. 적 엔티티 구현(전차 적, 보병 적) 및 이동
3. 벽 엔티티 구현 및 배치(카메라에 대한 상대 위치 고려 문제, 충돌 시 파괴 여부, 파괴 후 이미지, 벽 유형
고려하기(건물, 벽, 대전차 방호벽 등))
4. 충돌 로직 중 폭발 범주에 대한 적 데미지 구현

+++ LocalPVP2PShooting/src/core/CollisionManager.java
@@ -25,13 +25,13 @@ public class CollisionManager {
 
 		int aH = a.getHeight();
 		int aW = a.getWidth();
-		int aX = a.getWorldX();
-		int aY = a.getWorldY();
+		double aX = a.getWorldX();
+		double aY = a.getWorldY();
 
 		int bH = b.getHeight();
 		int bW = b.getWidth();
-		int bX = b.getWorldX();
-		int bY = b.getWorldY();
+		double bX = b.getWorldX();
+		double bY = b.getWorldY();
 
 		boolean overlapX = (aX < bX + bW) && (aX + aW > bX);
 		boolean overlapY = (aY < bY + bH) && (aY + aH > bY);
